Fix syslog messages lost after syslogd restarts
If syslogd is stopped and restarted while a process has the log open, the next
message that process produces will be logged to the console and not to the new
instance of syslogd. Further messages will be routed correctly, but not the
first one.

Based on original patch
Bug 3889 - syslog loses messages when syslogd restarted
https://bugs.busybox.net/show_bug.cgi?id=3889

function                                             old     new   delta
openlog                                               33      68     +35
__vsyslog                                            851     848      -3
openlog_intern                                       318     259     -59
------------------------------------------------------------------------------
(add/remove: 0/0 grow/shrink: 1/2 up/down: 35/-62)            Total: -27 bytes

Signed-off-by: Vladislav Grishenko <themiron@mail.ru>