ARCH_NATIVE_BIT should in the first place be passed when building utils for host
ARCH_NATIVE_BIT should in the first place be passed when building utils
for host (BUILD_CFLAGS-utils). It could also be passed when building utils
for target (CFLAGS-utils), but this is actually not necessary
as ARCH_NATIVE_BIT and __WORDSIZE are identical in this case.

Fixes 5ec4477584b69fe5f1410649b2eb3a63835fb649 (makes it actually work).

Signed-off-by: Eugene Rudoy <gene@freetz.org>