ARC: Disable obstack
Obstack was enabled in ARC defconfig to enable builds of native binutils and
GDB. This is not needed after earlier patch which removes invalid
_GNU_OBSTACK_INTERFACE_VERSION definition in cases when obstack is not
selected in uClibc configuration. With this change binutils and GDB would
use obstack implementation from their own source trees. In fact with recent
binutils it is not possible to build it using obstack implementation in
uClibc, because binutils/include/obstack.h started to use _obstack_free
function instead of obstack_free, but _obstack_free is not defined in
uClibc.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>
Cc: Vineet Gupta <Vineet.Gupta1@synopsys.com>
Cc: Alexey Brodkin <abrodkin@synopsys.com>