DNS: don't count search-path miss as a retry
Currently a miss on a search-path entry is counted as a retry.  This means that
users with more than (num_nameservers * retries) entries in their search path
list fail before trying all search paths.  Concretely, a single nameserver with
4 search paths will never try the 4th search because the default retry is 3.

The code doesn't currently retry a given nameserver in case of an error, so
retries is sort of meaningless (though there are some comments indicating it
might come).  This change only treats total failure of a nameserver (try next
server) as a retry.

Signed-off-by: Tim Hockin <thockin@google.com>