ldso: fix dlsym hang when reloading DSOs
It can happen under certain cases that the DSO had refcount 0,
but was already loaded. (NODELETE flag is set, or it is pulled
in via both NEEDED dependency and explicit dlopen()).

Add extra reference count for NODELETE objects, this will
ensure that the reference count never drops below one.

It is improved version of
 http://lists.busybox.net/pipermail/uclibc/2013-June/047826.html

Signed-off-by: Leonid Lisovskiy <lly.dev@gmail.com>