ARC: remove deprecated -msoft-float from CFLAGS
"-msoft-float" makes no sense for ARC because there's no such thing
as "-mhard-float" on ARC. Instead we use our own "-mfpu=XXX" option
when a particular floating-point related HW feature is enabled in
the ARC core.

We used to live with that phony option for quite some time but
with migration to newer GCC following warning now appears:
-------------------->8-----------------
arc-linux-gcc: warning: ‘-msoft-float’ is deprecated
cc1: warning: ‘-msoft-float’ is deprecated [enabled by default]
-------------------->8-----------------

And that warning gets printed for each invocation of gcc with
the option in question, which makes compilation output barely
readable.

So we disable that phony option for ARC now.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>
Cc: Waldemar Brodkorb <wbx@uclibc-ng.org>
Cc: Vineet Gupta <vgupta@synopsys.com>
Cc: Anton Kolesov <akolesov@synopsys.com>