ARC: Fix max ULP for cosine test
Max ULP for "cos (M_PI_6l * 2.0) == 0.5" in "test-double" sub-test
is set to "0" for ARC. It causes "test-double" test failure.

I realized that this sub-test on ARM gets completely same results.
But it doesn't fails on ARM because max ULP is set to "1" for ARM.

I am wondering, if it was done for any special reason?

Otherwise, set max ULP for "cos (M_PI_6l * 2.0) == 0.5" test to
"1" instead of "0" for ARC.

Signed-off-by: Eugeniy Paltsev <Eugeniy.Paltsev@synopsys.com>