or1k: Fix compiling with PIC and latest binutils use PLT for __syscall_error
This symbol was causing a build failure with the new toolchain.  It
looks like it has always been wrong.

The main issue was checking for PIC rather than __PIC__.

Remove all PSEUDO_* macros and ther SYSCALL_ERROR_NAME macro as they are
not needed by uclibc-ng, they are used in glibc for building up syscalls
there, but not here.

Fixes error:
  /opt/shorne/software/or1k-linux/bin/../lib/gcc/or1k-linux/9.0.1/../../../../or1k-linux/bin/ld: libc/libc_so.a(or1k_clone.os): pc-relative relocation against dynamic symbol __syscall_error
  /opt/shorne/software/or1k-linux/bin/../lib/gcc/or1k-linux/9.0.1/../../../../or1k-linux/bin/ld: final link failed: bad value

Signed-off-by: Stafford Horne <shorne@gmail.com>