open: Add support for O_TMPFILE
Since Linux 3.11, O_TMPFILE allows to create unnamed files that can be
linked later on.  It is internally defined as (O_TMPFILE | O_DIRECTORY)
to make it fail on old kernels.

Copying definitions from glibc for O_TMPFILE is not enough to support
O_TMPFILE; The open() wrapper also need to pass the mode when the flag
contains O_TMPFILE, otherwise, it will pass mode 000 which will succeed
but yield unexpected results.

openat() is curiously not affected since it passes the mode
unconditionally..

Signed-off-by: Nicolas Cavallari <nicolas.cavallari@green-communications.fr>