arm: elf-fdpic.h: avoid void pointer subtraction
elf-fdpic.h is included by link.h. When a C++ program includes <link.h>,
we get the following build failure:

<...>/usr/include/bits/elf-fdpic.h: In function ‘void* __reloc_pointer(void*, const elf32_fdpic_loadmap*)’:
<...>/usr/include/bits/elf-fdpic.h:94:54: error: invalid use of ‘void’
   94 |       unsigned long offset = p - (void*)map->segs[c].p_vaddr;
      |                                                      ^~~~~~~

void pointer addition and subtraction is not allowed in C++ as it has
undetermined size, however in C with language extension it is possible
because sizeof void is treated as one byte.

This patch was previously applied to Blackfin, FR-V and C6x, but not
ARM.

Signed-off-by: Ben Wolsieffer <ben.wolsieffer@hefring.com>